+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.0.0
 * Use klutter_ui 1.1.0.
 * Use Klutter Gradle v2024.1.1.beta with support for protobuf.
+* Set iOS to 13.0 in Podfile and podspec.
 * Move post-build tasks in root/platform build.gradle.kts to gradle plugin.
 * Add kradle script (which replaces producer and consumer).
 * Add interactive cli mode.